% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearm_utils.R
\name{getFromREARMDB}
\alias{getFromREARMDB}
\title{Fetch the data from a rearm DB given a list with the necessery parameters}
\usage{
getFromREARMDB(dbConfig = NULL, where = NULL, label = NULL,
  decimatedMagnitude = 1, class = NULL, target = NULL, usemode = NULL)
}
\arguments{
\item{dbConfig}{cannot be null. It is a string with the absolute path of the json file that contains the information to access to a database.}

\item{where}{is a mysql `where` condition to fetch the data from the Monitoring table of the REARM database.}
}
\description{
\code{getFromREARMDB} is a function that was made to easily access the monitoring data
presented in the database. For example, the InMonitor datalogger generates 32 points per signal per minute,
virtually generating 32 rows with 6 measurements each minute, with a timestamp associated.
This functions simplifies the fetch of a set of rows by given, between other things, the timestamps.
}

